move truncated into slg
