+++ src/solve/slg/truncate.rs
